lib/upnp: Small refactor
1. Split into separate files
2. Use DialTimeout rather than Dial
3. Remove shouldDebug(), as it's no longer needed
4. Change localIp return argument to be an IP
5. Add Get{Local,External}IPAddress functions to IGD
